#52 refactor: getTimeDiff 함수를 유틸 함수로 분리
